+++ src/Types/IDEntityChassis.php
@@ -4,8 +4,6 @@ declare(strict_types = 1);
 
 namespace AvtoDev\IDEntity\Types;
 
-use AvtoDev\IDEntity\Helpers\Normalizer;
-use AvtoDev\IDEntity\Helpers\Transliterator;
 use AvtoDev\ExtendedLaravelValidator\Extensions\ChassisCodeValidatorExtension;
 
 class IDEntityChassis extends AbstractTypedIDEntity
@@ -33,26 +31,7 @@ class IDEntityChassis extends AbstractTypedIDEntity
      */
     public static function normalize($value): ?string
     {
-        try {
-            // Replace multiple whitespaces with one
-            $value = (string) \preg_replace('~\s+~u', ' ', \trim((string) $value));
-
-            // Normalize dash chars
-            $value = Normalizer::normalizeDashChar($value);
-
-            // Transliterate cyrillic chars with latin
-            $value = Transliterator::transliterateString(\mb_strtoupper($value, 'UTF-8'), true);
-
-            // Remove all chars except allowed
-            $value = (string) \preg_replace('~[^A-Z0-9\- ]~u', '', $value);
-
-            // Replace multiple dashes with one
-            $value = (string) \preg_replace('~-+~', '-', $value);
-
-            return $value;
-        } catch (\Throwable $e) {
-            return null;
-        }
+        return IDEntityBody::normalize($value);
     }
 
     /**