+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## Unreleased
+
+### Changed
+
+- Normalization of `VIN`, `BODY`, `CHASSIS`
+- Not used method marked as deprecated
+
 ## v5.10.0
 
 ### Added